# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  schemas.TenantCreate:
    properties:
      address:
        type: string
      cuit_pdv:
        type: string
      email:
        type: string
      identifier:
        type: string
      name:
        type: string
      phone:
        type: string
      plan_id:
        type: integer
    required:
    - address
    - cuit_pdv
    - email
    - identifier
    - name
    - phone
    - plan_id
    type: object
  schemas.TenantResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  schemas.TenantUserCreate:
    properties:
      tenant_create:
        $ref: '#/definitions/schemas.TenantCreate'
      user_create:
        $ref: '#/definitions/schemas.UserCreate'
    required:
    - tenant_create
    - user_create
    type: object
  schemas.UserCreate:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
info:
  contact: {}
  description: This is a api to app noa gestion
  termsOfService: http://swagger.io/terms/
  title: APP NOA Gestion
  version: "1.0"
paths:
...
  /api/v1/tenant/create:
    post:
      consumes:
      - application/json
      description: Tenant Create required auth token
      parameters:
      - description: UserID
        in: query
        name: user_id
        required: true
        type: integer
      - description: TenantCreate
        in: body
        name: TenantCreate
        required: true
        schema:
          $ref: '#/definitions/schemas.TenantCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Tenant creado con éxito
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Tenant Create
      tags:
      - Tenant
  /api/v1/tenant/create_tenant_user:
    post:
      consumes:
      - application/json
      description: Tenant Create required auth token
      parameters:
      - description: TenantUserCreate
        in: body
        name: TenantUserCreate
        required: true
        schema:
          $ref: '#/definitions/schemas.TenantUserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Tenant y Usuario creados con éxito
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Tenant Create
      tags:
      - Tenant
  /api/v1/tenant/get_all:
    get:
      consumes:
      - application/json
      description: Tenant GetAll required auth token
      produces:
      - application/json
      responses:
        "200":
          description: Tenants obtenidos con éxito
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.TenantResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Tenant GetAll
      tags:
      - Tenant
  /user/create:
    post:
      consumes:
      - application/json
      description: Creates a new user.
      parameters:
      - description: User information
        in: body
        name: userCreate
        required: true
        schema:
          $ref: '#/definitions/schemas.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Create User
      tags:
      - User
securityDefinitions:
  CookieAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"